Update SQL Instance sweeper to check common prefixes (#6934)
